Reorganisation of mock keys and add deployment deployment package will provide a singular object to provide to provisioning and orchestration systems like ansible, kubernetes, terraform
